Adapt the AnycubicTFT to Marlin 2.0
Added the hooks from the TFT to the right places in Marlin 2.0.
Also started to modify the behavior of the LCD code to better match
my behavior taste.

Some code still duplicates basic function of Marlin (f.e. checking
for too high or too low temperature). This logic should be moved to
the correspondig module (and we will then only be a consumer of the
event).